requirements.txt
Django==4.1
pycryptodome==3.15.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
requests==2.28.1
django-filter==22.1
psycopg2==2.9.4
coverage==6.5.0
jsonnet==0.18.0
django-nose==1.4.6
django-rest-swagger==2.2.0
selenium==4.7.2

django-unfold==0.16.0
locust==2.17.0
webdriver-manager==4.0.1
dj-database-url == 2.1.0
pynose == 1.4.8
whitenoise==6.5.0
gunicorn==21.2.0
locust==2.17.0
django-livereload-server
pynose == 1.4.8
django-admin-autocomplete-filter
social-auth-app-django
django-user-agents
